rag_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyYAML==6.0.2
httpx==0.27.2
pydantic==2.9.2
chromadb==0.5.5
sentence-transformers==2.2.2
numpy==1.26.4

